requirements.txt
langchain
openai
faiss-cpu
tiktoken
PyPDF2
python-dotenv